refactor(github): use githubFileForm with json forms
