+++ pages/docs/[...slug].tsx
@@ -18,14 +18,14 @@ import {
 } from '../../components/ui'
 import { InlineWysiwyg, InlineTextareaField } from '../../components/ui/inline'
 import { TinaIcon } from '../../components/logo'
-import { useLocalGithubMarkdownForm } from '../../utils/github/useLocalGithubMarkdownForm'
+import { useGithubMarkdownForm } from '../../utils/github/useGithubMarkdownForm'
 import { getDocProps } from '../../utils/docs/getDocProps'
 import OpenAuthoringSiteForm from '../../components/layout/OpenAuthoringSiteForm'
 import OpenAuthoringError from '../../open-authoring/OpenAuthoringError'
 
 function DocTemplate(props) {
   // Registers Tina Form
-  const [data, form] = useLocalGithubMarkdownForm(
+  const [data, form] = useGithubMarkdownForm(
     props.markdownFile,
     formOptions,
     props.sourceProviderConnection
